refactor: LogEntryFactoryをDomain層に移動してDDDアーキテクチャを改善
- LogEntryFactoryクラスをApplication層からDomain層へ移動
  - 移動元: AiDevTest1.Application/Factories/LogEntryFactory.cs
  - 移動先: AiDevTest1.Domain/Services/LogEntryFactory.cs
- ILogEntryFactoryインターフェースもDomain層へ移動
  - 移動元: AiDevTest1.Application/Interfaces/ILogEntryFactory.cs
  - 移動先: AiDevTest1.Domain/Interfaces/ILogEntryFactory.cs
- 名前空間を更新: AiDevTest1.Application.Factories → AiDevTest1.Domain.Services
- 全ての参照箇所を更新してusing文を活用したクリーンなコードに修正
- 空のFactoriesフォルダを削除
- 循環依存を解決してDDDの依存関係ルールに準拠

fix #50
