+++ AiDevTest1.Tests/LogEntryFactoryTests.cs
@@ -1,4 +1,4 @@
-using AiDevTest1.Application.Factories;
+using AiDevTest1.Domain.Services;
 using AiDevTest1.Domain.Models;
 using Xunit;
